+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useTransactionFilters';
 export * from './useTimePeriodFilter';
 export * from './useFocusTrap';
+export * from './useOnboarding';